requirements.txt
fastapi
uvicorn
pymongo
pandas
pypdf
sentence-transformers
faiss-cpu
google-generativeai
streamlit
python-dotenv